decomposed 'run_tests.py' into 'conditions.py', 'define.py', and 'run_tests.py'.
